fix(quay): improve code readability with hooks
